refactor: replace 8 per-spec geometry providers with GmlFeatureGeometryProvider<T>
Introduce GmlFeatureGeometryProvider<TFeature> in Core that implements
IFeatureGeometryProvider for any IGmlFeature. Delete the 8 identical
per-spec S{NNN}FeatureGeometryProvider classes and update all processors
and tests to use the generic version.
